(chore) Move watson directory under bot
* clean up require statements
