+++ bot/watson/tone.js
